mlflow_client/requirements/all.txt
boto3
darts==0.26.0
ipython==8.10.0
ipykernel==5.3.4
ipywidgets==7.5.1
jupyter==1.0.0
notebook==7.0.4
jupyterlab==4.0.3
pandas==2.1.1
ipython_genutils
numpy
mlflow[extra]==2.7.1
optuna==3.3.0
pytorch-lightning==2.1.0rc0
plotly-resampler
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
python-decouple
compress_pickle
xgboost==2.0.0
